Special handling for prefix replacement in ELF binaries.
This change avoids prefix replacements within the debug
sections of ELF binaries. There are path tables in the DWARF
sections of shared libraries that may include placeholders,
but doing naive regexp replacement in those sections causes
the debug information to become corrupted. Using the
resultant libraries doesn't appear to have any correctness
issues, but any interaction with the debug information (gdb,
valgrind) yelds bad stack information.

When doing binary replacement, try to open the binary as an
ELF file. If this succeeds, avoid replacing strings in any
.debug_ sections. This does mean that attempts by code to
step into source files that have not been prefix-replaced
will be incorrect. With a non-trivial amount of additional
effort, it may be possible to do prefix replacement in DWARF
directory tables and repack the structures so they work
properly after installation.

Check for sections of type 'SHT_NOBITS'. These have a size
and offset, but are actually indexes into the segment.
Trying to replace within the section.data() results in
replacing into subsequent sections. In the case of .bss,
that results in editing the debug sections we're trying to
avoid editing.

If the binary cannot be opened as an ELF file, a simple
regexp replacement is performed.